Made 'regenerate' its own api action, for testing purposes. Uploaded to appengine; looks good!
git-svn-id: https://cooper-redhen.googlecode.com/svn/trunk@352 c84bbe80-f878-e576-0e26-96495b094a5b

+++ frontend.py
@@ -80,18 +80,16 @@ class ApiHandler(webapp.RequestHandler):
         user = users.get_current_user()
         
         # Just for testing
-        '''
-        [db.delete(s) for s in backend.Spectrum.all(keys_only=True)]
-        [db.delete(s) for s in backend.Project.all(keys_only=True)]
-        [db.delete(s) for s in backend.GenericData.all(keys_only=True)]
-        memcache.flush_all()
-        import os
-        for s in os.listdir('infrared'):
-            if s[0]!='.': backend.add( open('infrared/'+s).read(), 'public', False)
-        for s in os.listdir('raman'):
-            if s[0]!='.': backend.add( open('raman/'+s).read(), 'public', False) 
-        #spectra = [ open('infrared/iodobenzene1.jdx').read() ]
-        '''
+        if action == 'regenerate':
+            [db.delete(s) for s in backend.Spectrum.all(keys_only=True)]
+            [db.delete(s) for s in backend.Project.all(keys_only=True)]
+            [db.delete(s) for s in backend.GenericData.all(keys_only=True)]
+            memcache.flush_all()
+            import os
+            for s in os.listdir('infrared'):
+                if s[0]!='.': backend.add( open('infrared/'+s).read(), 'public', False)
+            for s in os.listdir('raman'):
+                if s[0]!='.': backend.add( open('raman/'+s).read(), 'public', False) 
         
         # If not operating on the main project, try getting the private one.
         # But abort if target is not supposed to be a project.
